% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupCF.R
\name{groupCF}
\alias{groupCF}
\title{groupCF}
\usage{
groupCF(formula, data, ngroups = 4L, ...)
}
\arguments{
\item{formula}{A formula of the form \code{x ~ z | w1 + w2} where \code{x} is an endogenous variable in an instrumental variables model, being predicted here. \code{z} is one of the instruments, and \code{w1} and \code{w2}, etc., are covariates to be used to predict the effect of the instrument.}

\item{data}{A data.frame.}

\item{ngroups}{Number of quantiles to split tau into.}

\item{...}{Additional arguments to be passed to \code{grf::causal_forest()}.}
}
\description{
This is a wrapper for \code{grf::causal_forest()} which performs a causal forest estimation, pulls out the individual-level effect estimates (tau), and then returns a factor variable containing the quantiles of tau.
}
\details{
This function is called by magnifiedIV. You can also run Magnified IV by yourself (with any estimator) by running groupSearch, then adding the resulting group variable as a control in both IV stages and also interacted with the instrument. Or use \code{grf::causal_forest()} directly to estimate tau on the individual level, and use that to construct a sample weight.
}
\examples{
# Get data
data(CPS1985, package = 'AER')

# See how the effect of education on wage varies over all the variables in the data
# and then split the resulting individual coefficient estimates into 10 quantiles
# (note this example probably does not satisfy the theoretical unconfoundedness
# assumption of causal forest; this is just a code example)
edeffect <- groupCF(wage ~ education |
                      experience + age + ethnicity +
                      region + gender + occupation +
                      sector + union + married,
                    data = CPS1985, ngroups = 10)

table(edeffect)
}
