% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnifiedIV.R
\name{ownPersonalSSEsus}
\alias{ownPersonalSSEsus}
\title{Create Personalized F-Statistic}
\usage{
ownPersonalSSEsus(co, excl, mpred, sse0, n, k)
}
\arguments{
\item{co}{This is a vector of coefficients unique to one individual, containing only the coefficients that vary across the sample. For example, if the grouped regression model contains \code{z*group}, then \code{co} would be a single value containing an individual's \code{z} effect.}

\item{excl}{This is a data frame containing just the variable(s) over which the effects vary. So if the regression model contains \code{z*group}, this would be a data frame containing only \code{z}.}

\item{mpred}{This is a vector containing the regression model prediction *if all variables in \code{excl} were set to 0*.}

\item{sse0, n, k}{These are the sum of squared errors, the number of observations, and the number of coefficients in the original regression model.}
}
\description{
This function creates an F-statistic of the personalized kind described in Huntington-Klein (2019).
}
\details{
In particular, it calculates (N-K)*Var(x-hat | individual coefficients)/Var(x-hat).

In Magnified IV this is raised to the power p to be used as a regression weight. This funciton is largely used internally for \code{magnifiedIV()} but it is exported here for use in case you'd like to use it to create your own weights. It is perhaps not as user-friendly as it could be, but it is largely an internal function.
}
\examples{

df <- data.frame(w1 = rnorm(1000),
                 w2 = rnorm(1000),
                 e1 = rnorm(1000),
                 e2 = rnorm(1000),
                 z = rnorm(1000),
                 groups = factor(floor(0:999/100)))
df$x <- df$w1+df$w2+df$z+df$e1

fsmodel <- lm(x ~ z*groups + w1 + w2, data = df)

sse0 <- sum((fsmodel$residuals)^2)

indivfx <- factorPull(fsmodel,
                      data = df,
                      factor = 'groups',
                      interaction = 'z',
                      addterm = 'z')

# Create prediction for each observation with just the coefficients that are the
# same for everyone
mpred <- rowSums(t(coef(fsmodel)[c('(Intercept)','w1','w2')]*
                     t(model.matrix(fsmodel)[,c('(Intercept)','w1','w2')])))
# Get a data frame of just the variables with effects that vary
excl <- data.frame(z = df[['z']])

# Get the relevant N and K
n <- nrow(df)
k <- length(fsmodel$coefficients)

# And finally produce that individualized F statistic
indiv.F <- sapply(indivfx, function(x)
  ownPersonalSSEsus(x, excl, mpred, sse0, n, k))
}
